pine-ai/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# Database
sqlalchemy==2.0.40
asyncpg==0.31.0
alembic==1.15.2

# YouTube  (1.2.4 is the first release supporting Python 3.14)
youtube-transcript-api==1.2.4

# AI / ML
sentence-transformers==5.4.1
faiss-cpu==1.13.2
numpy==2.2.5

# HTTP client
httpx==0.28.1

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
anyio==4.9.0

# Logging / utilities
python-dotenv==1.1.0
